requirements.txt
tensorflow==2.12.0
keras-tuner==1.3.4
keras-tcn==3.5.0
pandas==1.5.3
scikit-learn==1.2.2
torch==1.12.1